Create _* procedures for handling configs
Use _* procedures as a backup for a possible GUI feature:
When opening a node Configure menu, create a copy of the node
configuration and edit it without modifying the real configuration
until Apply is pressed. We can then parse the new configuration and
apply it as it should be, especially in exec mode (watching out for the
order of executing things).