Disable node reconfig for some cases
When custom config is enabled or when auto default routes is disabled on
node, do not automatically reconfigure them. This would usually happen
when connecting/disconnecting a router/extnat node in the same subnet or
when changing some router/extnats interface IP address.

Also, do not reconfigure if custom config is selected and the user
toggles auto default routes in node config.